Exercise 1 Finished
Clone the GitHub repo, create your MPLAB x project and add all the .c and .h files to your project (the timers.c/h files are not needed until exercise 2).
Complete the interrupt initialisation function and ISR to toggle the LED on RH3 when the LDR goes from light to dark. Remember to call the necessary init functions from your main and set the TRIS register for any outputs.
There should be nothing in the main while(1) loop. To conserve power, call the Sleep() function from the main while(1) loop to put the chip to sleep when the ISR finishes. How can you check that the chip has entered sleep mode?